somewhat working wayland screen
