+++ src/core/screen/linux/wayland.rs
@@ -1,4 +1,4 @@
-use crate::core::CaptureableScreen;
+use crate::core::{CaptureableScreen, keyboard::Keyboard};
 #[cfg(not(feature = "lite"))]
 use crate::errors::AutoGuiError;
 use crate::imgtools;
@@ -8,7 +8,7 @@ use crossbeam_channel::Receiver;
 #[cfg(not(feature = "lite"))]
 use image::{GrayImage, ImageBuffer, Luma, Rgba, RgbaImage};
 use std::{
-    sync::{mpsc::RecvTimeoutError, Arc, RwLock},
+    sync::{Arc, RwLock, mpsc::RecvTimeoutError},
     thread,
     time::Duration,
 };
@@ -17,7 +17,7 @@ use waycap_rs::{Capture, RgbaImageEncoder};
 pub struct WaylandScreen {
     capture: Capture<RgbaImageEncoder>,
     latest_img: Receiver<RgbaImage>,
-    img: Option<RgbaImage>,
+    img: RgbaImage,
 }
 
 impl CaptureableScreen for WaylandScreen {
@@ -25,27 +25,31 @@ impl CaptureableScreen for WaylandScreen {
     where
         Self: std::marker::Sized,
     {
-        let mut cap = Capture::new_with_encoder(RgbaImageEncoder::default(), false, 30).unwrap();
+        let mut cap = Capture::new_with_encoder(RgbaImageEncoder::default(), false, 60).unwrap();
         let recv = cap.get_output();
         let recv_thread = recv.clone();
-        thread::spawn(move || loop {
-            // drain channel to get latest message
-            while recv_thread.len() > 1 {
-                recv_thread.try_recv().ok();
+        thread::spawn(move || {
+            loop {
+                // drain channel to get latest message
+                while recv_thread.len() > 1 {
+                    recv_thread.try_recv().ok();
+                }
+                // avoid busy waiting
+                std::thread::yield_now();
             }
-            // avoid busy waiting
-            std::thread::yield_now();
         });
-        Ok(WaylandScreen {
+        let img = recv.recv().expect("Could not get first image");
+        let screen = WaylandScreen {
             capture: cap,
             latest_img: recv,
-            img: None,
-        })
+            img,
+        };
+        screen.dimension();
+        Ok(screen)
     }
     /// returns screen dimensions. All monitors included
     fn dimension(&self) -> (i32, i32) {
-        self.get_img()
-            .map_or((0, 0), |img| (img.width() as i32, img.height() as i32))
+        (self.img.width() as i32, self.img.height() as i32)
     }
 
     fn destroy(&mut self) {
@@ -60,8 +64,8 @@ impl CaptureableScreen for WaylandScreen {
         &mut self,
         (x, y, width, height): (u32, u32, u32, u32),
     ) -> Result<RgbaImage, AutoGuiError> {
-        self.get_img_mut()
-            .map(|img| image::imageops::crop(img, x, y, width, height).to_image())
+        self.capture_screen()?;
+        Ok(image::imageops::crop(&mut self.img, x, y, width, height).to_image())
     }
 
     /// executes convert_bitmap_to_grayscale, meaning it converts Vector of values to grayscale and crops the image
@@ -71,54 +75,42 @@ impl CaptureableScreen for WaylandScreen {
         &mut self,
         &(x, y, width, height): &(u32, u32, u32, u32),
     ) -> Result<GrayImage, AutoGuiError> {
-        self.get_img_mut().and_then(|img| {
-            convert_rgba_to_bw(imgtools::cut_screen_region(x, y, width, height, img))
-        })
+        self.capture_screen()?;
+        convert_rgba_to_bw(imgtools::cut_screen_region(x, y, width, height, &self.img))
     }
     #[cfg(not(feature = "lite"))]
     /// captures and saves screenshot of monitors
     fn grab_screenshot(&mut self, image_path: &str) -> Result<(), AutoGuiError> {
         self.capture_screen()?;
-        self.get_img()?.save(image_path).map_err(Into::into)
+        self.img.save(image_path).map_err(Into::into)
     }
     #[cfg(not(feature = "lite"))]
     /// first order capture screen function. it captures screen image and stores it as vector in self.pixel_data
     fn capture_screen(&mut self) -> Result<(), AutoGuiError> {
-        if self.img.is_none() | !self.latest_img.is_empty() {
-            self.img = self.latest_img.try_recv().ok();
+        if !self.latest_img.is_empty() {
+            self.img = self
+                .latest_img
+                .try_recv()
+                .expect("reciever was not empty, but could not get a image");
         }
         Ok(())
     }
     #[cfg(not(feature = "lite"))]
     /// convert vector to Luma Imagebuffer
     fn convert_bitmap_to_grayscale(&self) -> Result<GrayImage, AutoGuiError> {
-        self.get_img()
-            .and_then(|img| convert_rgba_to_bw(img.clone()))
+        convert_rgba_to_bw(self.img.clone())
     }
     #[cfg(not(feature = "lite"))]
     /// convert vector to RGBA ImageBuffer
     fn convert_bitmap_to_rgba(&self) -> Result<RgbaImage, AutoGuiError> {
-        self.get_img().cloned()
+        Ok(self.img.clone())
     }
 
     fn create_keyboard(&mut self) -> crate::core::keyboard::Keyboard {
-        todo!()
+        Keyboard::Wayland
     }
 
     fn create_mouse(&mut self) -> crate::core::mouse::Mouse {
-        todo!()
-    }
-}
-
-impl WaylandScreen {
-    fn get_img(&self) -> Result<&RgbaImage, AutoGuiError> {
-        self.img.as_ref().ok_or(AutoGuiError::OSFailure(
-            "No image from wayland capture yet".to_string(),
-        ))
-    }
-    fn get_img_mut(&mut self) -> Result<&mut RgbaImage, AutoGuiError> {
-        self.img.as_mut().ok_or(AutoGuiError::OSFailure(
-            "No image from wayland capture yet".to_string(),
-        ))
+        crate::core::mouse::Mouse::Wayland
     }
 }
